+++ src/form/common/default-validation-conditionals.ts
@@ -317,7 +317,7 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['mother.iD']
+        parameters: ['mother.motherNationalId']
       }
     ]
   } else if (configCase === 'mother') {
@@ -328,7 +328,7 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['father.iD']
+        parameters: ['father.fatherNationalId']
       }
     ]
   } else if (configCase === 'deceased') {
@@ -339,7 +339,7 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['informant.informantID']
+        parameters: ['informant.informantNationalId']
       }
     ]
   } else if (configCase === 'groom') {
@@ -350,7 +350,7 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['bride.iD']
+        parameters: ['bride.brideNationalId']
       }
     ]
   } else if (configCase === 'bride') {
@@ -361,7 +361,7 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['groom.iD']
+        parameters: ['groom.groomNationalId']
       }
     ]
   } else {
@@ -373,23 +373,23 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['deceased.deceasedID']
+        parameters: ['deceased.deceasedNationalId']
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['mother.iD']
+        parameters: ['mother.motherNationalId']
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['father.iD']
+        parameters: ['father.fatherNationalId']
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['groom.iD']
+        parameters: ['groom.groomNationalId']
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['bride.iD']
+        parameters: ['bride.brideNationalId']
       }
     ]
   }